+++ frontend/src/app/components/containers/complaints/details/complaint-details-create.tsx
@@ -37,7 +37,7 @@ import { DismissToast, TOAST_POSITION, ToggleError, ToggleInformation } from "@c
 import { useNavigate } from "react-router-dom";
 import { Attachments } from "@components/common/attachments-carousel";
 import { COMSObject } from "@apptypes/coms/object";
-import { handleAddAttachments, handleDeleteAttachments, handlePersistAttachments } from "@common/attachment-utils";
+import { handleAddAttachments, handleDeleteAttachments } from "@common/attachment-utils";
 import { uploadAttachmentsWithProgress } from "@common/attachment-upload-helper";
 
 import { WildlifeComplaint } from "@apptypes/app/complaints/wildlife-complaint";
@@ -56,7 +56,6 @@ import { ParkSelect } from "@/app/components/common/park-select";
 import { isValidEmail } from "@/app/common/validate-email";
 import { AgencyType } from "@/app/types/app/agency-types";
 import { ValidationDatePicker } from "@/app/common/validation-date-picker";
-import { Id } from "react-toastify";
 import { attachmentUploadComplete$ } from "@/app/types/events/attachment-events";
 import useUnsavedChangesWarning, { useFormDirtyState } from "@/app/hooks/use-unsaved-changes-warning";
