+++ frontend/src/app/components/containers/investigations/details/investigation-contravention/contravention-details-form.tsx
@@ -20,7 +20,7 @@ import { ValidationDatePicker } from "@/app/common/validation-date-picker";
 import { useAppSelector } from "@/app/hooks/hooks";
 import { selectCommunityCodeDropdown } from "@/app/store/reducers/code-table";
 import { format } from "date-fns";
-import { parseUTCDateTimeToLocal } from "@/app/common/methods";
+import { parseUTCDateTimeToLocal } from "@/app/common/date-utils";
 
 export interface ContraventionDetailsFormValues {
   contraventionDate: string;
