+++ frontend/src/app/components/containers/investigations/details/investigation-diary-dates/diary-date-row.tsx
@@ -1,13 +1,12 @@
 import { FC } from "react";
 import { DiaryDate } from "@/generated/graphql";
-import { parseUTCDateTimeToLocal } from "@common/methods";
 import { useAppSelector } from "@/app/hooks/hooks";
 import { selectOfficerByAppUserGuid } from "@/app/store/reducers/officer";
 import { useNavigate, useParams } from "react-router-dom";
 import { InvestigationParams } from "@/app/components/containers/investigations/details/investigation-details";
 import { EditButton } from "@components/common/comp-table-edit-column";
 import { useInvestigationReadOnly } from "../../hooks/use-investigation-read-only";
-import { formatDateObjectAsString, parseUTCTimestampToLocal } from "@/app/common/date-utils";
+import { formatDateObjectAsString, parseUTCTimestampToLocal, parseUTCDateTimeToLocal } from "@/app/common/date-utils";
 
 interface DiaryDateRowProps {
   diaryDate: DiaryDate;
