+++ frontend/src/app/components/containers/parties/list/party-list.tsx
@@ -5,7 +5,8 @@ import { CompColumn } from "@/app/types/app/comp-tables";
 import { usePartySearch } from "../hooks/use-party-search";
 import { SORT_TYPES } from "@constants/sort-direction";
 import { PartyTypeCodes } from "@/app/constants/party-types";
-import { calculateAgeYears, parseUTCDateTimeToLocal } from "@common/methods";
+import { calculateAgeYears } from "@common/methods";
+import { parseUTCDateTimeToLocal } from "@/app/common/date-utils";
 import { ContactMethods } from "@/app/constants/contact-methods";
 import { BusinessIdentifiers } from "@/app/constants/business-identifiers";
 import { formatPhoneNumber } from "react-phone-number-input/input";
