+++ frontend/src/app/components/containers/parties/match/party-match-card.tsx
@@ -3,7 +3,8 @@ import { FC } from "react";
 import { Button, Card } from "react-bootstrap";
 import { useAppSelector } from "@hooks/hooks";
 import { selectGenderDropdown } from "@/app/store/reducers/code-table";
-import { calculateAgeYears, parseUTCDateTimeToLocal } from "@/app/common/methods";
+import { calculateAgeYears } from "@/app/common/methods";
+import { parseUTCDateTimeToLocal } from "@/app/common/date-utils";
 import { ContactMethods } from "@/app/constants/contact-methods";
 import { Address, Alias, BusinessIdentifier, ContactMethod, Party } from "@/generated/graphql";
 import { formatPhoneNumber } from "react-phone-number-input";
